Move overview info form into a component
