+++ app/components/overview-info-form/template.hbs
@@ -1,0 +1,7 @@
+<p>
+    Welcome! We are interested in the situations people experience and what they do in them. You will describe a
+    situation you experienced recently and what you did in that situation. You will also be asked some questions about
+    your attitudes and values. Based on these responses, when you have completed the study, you will receive
+    individualized information about your personality that we hope you will find interesting.
+</p>
+{{dynamic-form schema=formSchema formActions=formActions}}
