+++ app/templates/survey.hbs
@@ -1,1 +1,2 @@
-{{dynamic-form schema=model}}
+<h3>International Situations Project</h3>
+{{overview-info-form}}
