+++ tests/integration/components/overview-info-form/component-test.js
@@ -1,0 +1,24 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('overview-info-form', 'Integration | Component | overview info form', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{overview-info-form}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#overview-info-form}}
+      template block text
+    {{/overview-info-form}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
